refactor: break out the 9P bridge to its own @endo/9p-server package
The 9P-over-UDS server in claude-container's `src/9p/` +
`fs-bridge-9p.js` was already independent of the rest of the
container stack — it consumed any `@endo/remote-fs` `Filesystem`
cap and served it to anything speaking 9P. Splitting it out lets
other consumers depend on the bridge directly (the smoke-boot
host responder, third-party 9P servers, future protocol bridges)
without pulling in claude-container's factory + orchestrator
client.

Layout:

  packages/9p-server/
    README.md
    package.json
    src/
      index.js          re-exports
      fs-bridge.js      makeFsBridge9p (was fs-bridge-9p.js)
      server.js         9P2000.L state machine (was 9p/server.js)
      wire.js           framing + LE primitives (was 9p/wire.js)
      types.js          T, QT, errno, mode constants (was 9p/types.js)
    test/
      wire.test.js      (was claude-container/test/9p-wire.test.js)
      server.test.js    (was claude-container/test/9p-server.test.js)

Mechanical moves via `git mv`. The internal import in
fs-bridge.js drops the './9p/' prefix (now './server.js'). Test
imports drop the same. No code logic changed.

Downstream import sites updated:

  - claude-container/src/fs-bridge-module.js imports from
    @endo/9p-server (was the local fs-bridge-9p.js)
  - claude-container/test/factory-live.test.js imports
    wire/types from @endo/9p-server/src/...
  - claude-orch/scripts/smoke-boot-host.js imports from
    @endo/9p-server (was @endo/claude-container/src/fs-bridge-9p.js)

Package.json wiring:

  - claude-container: drops `./src/fs-bridge-9p.js` export entry,
    adds @endo/9p-server runtime dep.
  - claude-orch devDeps: swap @endo/claude-container for
    @endo/9p-server (smoke-boot-host needs the bridge, not the
    rest of claude-container; removes the workspace cycle).

Docs updated: README, ENDO-INTEGRATION §6.2 + §9-R1 + layout
section all point at @endo/9p-server instead of the old
in-tree paths.

Tests:
  - @endo/9p-server: 18 (4 framing + 14 protocol)
  - @endo/claude-container: 16 (was 34; 18 protocol tests moved out)
  - @endo/claude-orch: 39
  - @endo/remote-fs: 101
  → 174 total, all green. Lint + build green.
